# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SynCloud API",
    "version": "v1",
    "description": "The public SynCloud controller API. The dashboard, synctl and CI all use these endpoints (plan \u00a75.1). Authenticate with the dashboard session cookie, a personal access token (Bearer), or a signed access-key request (SYN1-HMAC-SHA256, plan \u00a77.1)."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "session": []
    },
    {
      "bearerToken": []
    },
    {
      "accessKey": []
    }
  ],
  "paths": {
...
    "/api/v1/projects/{project}/security-groups": {
      "get": {
        "operationId": "listSecurityGroups",
        "tags": [
          "network"
        ],
        "summary": "A project's security groups",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SecurityGroup"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "operationId": "createSecurityGroup",
        "tags": [
          "network"
        ],
        "summary": "Create a security group",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SecurityGroupRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecurityGroup"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/projects/{project}/security-groups/{group}": {
      "get": {
        "operationId": "getSecurityGroup",
        "tags": [
          "network"
        ],
        "summary": "A security group with rule hit counters",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "group",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecurityGroup"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "put": {
        "operationId": "updateSecurityGroup",
        "tags": [
          "network"
        ],
        "summary": "Replace a security group's rules and attachments",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "group",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SecurityGroupRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecurityGroup"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "deleteSecurityGroup",
        "tags": [
          "network"
        ],
        "summary": "Delete a security group (not the default one, nor one other rules name)",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "group",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "syncloud_session"
      },
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Personal access token (syn_pat_\u2026)"
      },
      "accessKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "SYN1-HMAC-SHA256 Credential=<access key id>, Signature=<hex>, with X-Syn-Date and X-Syn-Content-Sha256 headers"
      }
    },
...
    "responses": {
      "Error": {
        "description": "Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "SecurityRule": {
        "type": "object",
        "properties": {
          "protocol": {
            "type": "string",
            "enum": [
              "tcp",
              "udp",
              "icmp",
              "any"
            ]
          },
          "ports": {
            "type": "string"
          },
          "peers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "any, cluster, IPv4/CIDR, group:[project/]name, service:[project/]env/name, environment:self|[project/]env, project:self|name"
          },
          "description": {
            "type": "string"
          },
          "packets": {
            "type": "integer",
            "readOnly": true
          },
          "bytes": {
            "type": "integer",
            "readOnly": true
          }
        }
      },
      "SecurityGroup": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "project": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "default": {
            "type": "boolean"
          },
          "inbound": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SecurityRule"
            }
          },
          "outbound": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SecurityRule"
            }
          },
          "services": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "attached services, ENV/NAME"
          },
          "members": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "every service using the group"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SecurityGroupRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "inbound": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SecurityRule"
            }
          },
          "outbound": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SecurityRule"
            }
          },
          "services": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  }
}